final_pjt_back/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
dj-rest-auth==6.0.0
Django==4.2.13
django-allauth==0.63.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.1
exceptiongroup==1.2.1
executing==2.0.1
fonttools==4.51.0
idna==3.7
importlib_resources==6.4.0
ipython==8.18.1
jedi==0.19.1
kiwisolver==1.4.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
zipp==3.18.1
